iptables: move custom chains logic into custom_chain.go
[ upstream commit 597ad32c1c58af210953bb3f80912ab1d810d8db ]

This commit moves all the logic related to custom chains into its own
custom_chain.go file. No functional changes are introduced.
